output get request for todolists

+++ app/src/main/java/com/volohov/todo/api/ApiDataModel.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 object  ApiDataModel {
     data class ToDoLists(
         @SerializedName("completed") val isCompleted: Boolean,
-        @SerializedName("completion_progress") val completionProgress: Int,
+        @SerializedName("completion_progress") val completionProgress: Double,
         @SerializedName("name") val name: String
     )
 }
